Implement ResponseError::FailedToStoreData(Vec<u8>)
